+++ js/mocks/photos.js
